Allow choosing a default target when rebooting

+++ hypr/scripts/power_menu.sh
@@ -7,6 +7,16 @@ case $(echo -e "Lock\nShutdown\nRestart" | rofi -dmenu -config ~/.config/rofi/mi
         poweroff
         ;;
     Restart)
-        reboot
+        if [ ! -e /boot/loader/entries ]; then
+            # we aren't using systemd-boot and i didn't write any other targeting logic
+            reboot;
+        fi;
+        TARGET=$(
+                for entry in $(ls /boot/loader/entries);
+                do
+                    echo $entry
+                done
+            ) | rofi -dmenu -config ~/.config/rofi/midscreen.rasi -p " "
+            systemctl reboot --boot-loader-entry=$TARGET
         ;;
 esac
